+++ src/main/java/com/stakevault/betting/stats/domain/model/BetType.java
@@ -1,0 +1,10 @@
+package com.stakevault.betting.stats.domain.model;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public enum BetType {
+	@JsonProperty("pre")
+	PRE,
+	@JsonProperty("live")
+	LIVE
+}