+++ src/main/java/com/stakevault/betting/stats/domain/port/in/CalculateMetricsUseCase.java
@@ -3,6 +3,7 @@ package com.stakevault.betting.stats.domain.port.in;
 import java.util.List;
 
 import com.stakevault.betting.stats.domain.model.BetMetrics;
+import com.stakevault.betting.stats.domain.model.DailyBetMetrics;
 import com.stakevault.betting.stats.domain.model.MonthlyBetMetrics;
 import com.stakevault.betting.stats.domain.model.SegmentedBetMetrics;
 import com.stakevault.betting.stats.domain.model.StatisticsFilter;
@@ -17,5 +18,13 @@ public interface CalculateMetricsUseCase {
 
 	List<SegmentedBetMetrics> calculateByBettingHouse(StatisticsFilter filter);
 
+	List<SegmentedBetMetrics> calculateByLeague(StatisticsFilter filter);
+
+	List<SegmentedBetMetrics> calculateByTipster(StatisticsFilter filter);
+
 	List<MonthlyBetMetrics> calculateMonthly(StatisticsFilter filter);
+
+	List<SegmentedBetMetrics> calculateByBetType(StatisticsFilter filter);
+
+	List<DailyBetMetrics> calculateDaily(StatisticsFilter filter);
 }